@@ -1,6 +1,6 @@
 module golang.stackrox.io/grpc-http1
 
-go 1.14
+go 1.16
 
 require (
 	github.com/golang/glog v1.0.0
